+++ practica1/src/vertex.hpp
@@ -8,33 +8,33 @@
 class Vertex {
 
     private:
-        double coordinates[4];
+        float coordinates[4];
 
     public:
         Vertex() { coordinates[0] = 0; coordinates[1] = 0; coordinates[2] = 0; coordinates[3] = 1; };
-        Vertex( const double x, const double y, const double z );
+        Vertex( const float x, const float y, const float z );
         Vertex( const Vertex& otherVertex );
-        double getX() const;
-        double getY() const;
-        double getZ() const;
-        double getH() const { return coordinates[3]; };
-        double get(int pos) const { return coordinates[pos]; };
-        void setX( double x ) { coordinates[0] = x; };
-        void setY( double y ) { coordinates[1] = y; };
-        void setZ( double z ) { coordinates[2] = z; };
-        void setH( double h ) { coordinates[3] = h; };
-        void set( int pos, double val );
-        void set (double x, double y, double z){ coordinates[0] = x; coordinates[1] = y; coordinates[2] = z;};
-        double module();
+        float getX() const;
+        float getY() const;
+        float getZ() const;
+        float getH() const { return coordinates[3]; };
+        float get(int pos) const { return coordinates[pos]; };
+        void setX( float x ) { coordinates[0] = x; };
+        void setY( float y ) { coordinates[1] = y; };
+        void setZ( float z ) { coordinates[2] = z; };
+        void setH( float h ) { coordinates[3] = h; };
+        void set( int pos, float val );
+        void set (float x, float y, float z){ coordinates[0] = x; coordinates[1] = y; coordinates[2] = z;};
+        float module();
         void normalize();
         bool operator==( const Vertex &otherVertex );
-        Vertex operator*( const double matrix[4][4] );
+        Vertex operator*( const float matrix[4][4] );
         Vertex operator*( const Vertex &otherVertex );
         void operator=( const Vertex &otherVertex );
         Vertex operator-( const Vertex &otherVertex );
         Vertex operator+( const Vertex &otherVertex );
-        Vertex operator/( const double &val );
-        Vertex operator*( const double &val );
+        Vertex operator/( const float &val );
+        Vertex operator*( const float &val );
 
 };
